# api/openapi.yaml
...
openapi: "3.0.2"
info:
  title: "Boot Orchestration Service"
  version: "1.3.0"
  description: |
    The Boot Orchestration Service (BOS) applies 1 of 4 possible
    actions (boot, reboot, reconfigure, shutdown) to a list of compute nodes
    using a specified image (if applicable), with specified kernel parameters,
    over a specified network. Currently, only one network option is available.
    After boot or reboot, nodes can be reconfigured.


    The default content type for the BOS API is "application/json". Unsuccessful
    API calls return a content type of "application/problem+json" as per RFC 7807.

    ## Resources


    ### /sessiontemplate

    A session template controls the activity of BOS; it is a collection of one or more boot
    sets, combined with partition information, and
    configuration detail.

    A boot set defines a list of nodes, the image you want to boot/reboot the nodes with,
    kernel parameters to use to boot the nodes, and the network to boot the nodes over.

    ### /session

    A BOS session applies the action to the nodes in the session
    template.

    ## Workflow


    ### Create a New Session

    #### GET /sessiontemplate

    List available session templates.
    Note the *name* which uniquely identifies each session template.
    This value can be used to create a new session later,
    if specified in the request body of POST /session.

    #### POST /sessiontemplate

...
    session template.
...

    #### POST /session

...

...

...


...

...

...


...


...


...
servers:
...
components:
  schemas:
...
    Link:
      description: Link to other resources
      type: object
      properties:
        rel:
          type: string
        href:
          type: string
      additionalProperties: false
...
    GenericMetadata:
      type: object
      description: |
        The status metadata
      properties:
        start_time:
          type: string
          description: |
            The start time
          example: "2020-04-24T12:00"
        stop_time:
          type: string
          description: |
            The stop time
          example: "2020-04-24T12:00"
        complete:
          type: boolean
          description: |
            Is the object's status complete
          example: true
        in_progress:
          type: boolean
          description: |
            Is the object still doing something
          example: false
        error_count:
          type: integer
          description: |
            How many errors were encountered
          example: 0
      additionalProperties: false
...
    PhaseStatus:
      type: object
      description: |
        The phase's status. It is a list of all of the nodes in the phase and
        what category those nodes fall into within the phase.
      
        ## Link Relationships
        
        * self : The session object

      properties:
        name:
          type: string
          description: |
            Name of the Phase 
          example: "Boot"
          pattern: '^(?i)boot|configure|shutdown$'
        metadata:
          $ref: '#/components/schemas/GenericMetadata'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/PhaseCategoryStatus'
        errors:
          $ref: '#/components/schemas/NodeErrorsList'
    BootSetStatus:
      type: object
      description: |
        The status for a Boot Set. It as a list of the phase statuses for the Boot Set.
         
        ## Link Relationships
        
        * self : The session object
        * phase : A phase of the boot set

      properties:
        name:
          type: string
          minLength: 1
          description: Name of the Boot Set
          example: "Boot-Set"
        session:
          type: string
          description: Session ID
          example: "Session-ID"
        metadata:
          $ref: '#/components/schemas/GenericMetadata'
        phases:
          type: array
          items:
            $ref: '#/components/schemas/PhaseStatus'
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    SessionStatus:
      type: object
      description: |
        The status for a Boot Session. It is a list of all of the Boot Set Statuses in the session.
        ## Link Relationships
        
        * self : The session object
        * boot sets: URL to access the Boot Set status

      properties:
        metadata:
          $ref: '#/components/schemas/GenericMetadata'
        boot_sets:
           description: |
             The boot sets in the Session
           type: array
           items:
             type: string
           minItems: 1
        id:
          type: string
          description: Session ID
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    SessionTemplate:
      type: object
      description: |
        A Session Template object represents a collection of resources and metadata.
        A session template is used to create a Session which when combined with an
        action (i.e. boot, reconfigure, reboot, shutdown) will create a K8s BOA job
        to complete the required tasks for the operation.

        A Session Template can be created from a JSON structure.  It will return
        a SessionTemplate name if successful.
        This name is required when creating a Session.

        ## Link Relationships

        * self : The session object
      properties:
        templateUrl:
          type: string
          description: |
            The URL to the resource providing the session template data.
            Specify either a templateURL, or the other session
            template parameters (excluding templateBody).
        templateBody:
          type: string
          description: |
            Do not use. To be removed.
        name:
          type: string
          description: Name of the SessionTemplate. The length of the name is restricted to 45 characters.
          example: "cle-1.0.0"
          # These validation parameters are restricted by Kubernetes naming conventions.
          minLength: 1
          maxLength: 45
          pattern: "[a-z0-9]([-a-z0-9]*[a-z0-9])?(\\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*"
        description:
          type: string
          description: |
            An optional description for the session template.
        cfs_url:
          type: string
          description: |
            The url for the repository providing the configuration. DEPRECATED
        cfs_branch:
          type: string
          description: |
            The name of the branch containing the configuration that you want to
            apply to the nodes.  DEPRECATED.
        enable_cfs:
          type: boolean
          description: |
            Whether to enable the Configuration Framework Service (CFS).
            Choices: true/false
          default: true
        cfs:
          $ref: '#/components/schemas/CfsParameters'
        partition:
          type: string
          description: |
            The machine partition to operate on.
        boot_sets:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/BootSet'
        links:
          type: array
          readOnly: true
          items:
            $ref: '#/components/schemas/Link'
      required: [name]
      additionalProperties: false
    Session:
      description: |
        A Session object

        ## Link Relationships

        * self : The session object
      type: object
      properties:
        operation:
          type: string
          description: >
            A Session represents an operation on a SessionTemplate.
            The creation of a session effectively results in the creation
            of a K8s Boot Orchestration Agent (BOA) job to perform the
            duties required to complete the operation.

            Operation -- An operation to perform on nodes in this session.

                Boot         Boot nodes that are off.

                Configure    Reconfigure the nodes using the Configuration Framework
                             Service (CFS).

                Reboot       Gracefully power down nodes that are on and then power
                             them back up.

                Shutdown     Gracefully power down nodes that are on.

          pattern: '^(?i)boot|configure|reboot|shutdown$'
        templateUuid:
          type: string
          description: The name of the Session Template
          example: "my-session-template"
          format: string
        limit:
          type: string
          description: >
            A comma seperated of nodes, groups or roles to which the session
            will be limited. Components are treated as OR operations unless
            preceded by "&" for AND or "!" for NOT.
        links:
          type: array
          readOnly: true
          items:
            $ref: '#/components/schemas/Link'
      required: [operation, templateUuid]
      additionalProperties: false
...
    NodeChangeList:
      type: object
      description: |
        The information used to update the status of a node list. It moves nodes from
        one category to another within a phase.
      properties:
        phase:
          type: string
          example: "Boot"
        source:
          type: string
          example: "in_progress"
        destination:
          type: string
          example: "Succeeded"
        node_list:
          $ref: '#/components/schemas/NodeList'
      additionalProperties: false
      required: [phase, source, destination, node_list]
...
    UpdateRequestNodeChangeList:
      description: |
        This is the payload sent during an update request. It contains 
        updates to which categories nodes are in.
      type: array
      items:
        type: object
        properties:
          update_type:
            description: The type of update data 
            pattern: "NodeChangeList"
            type: string
          phase:
            description: |
              The phase that this data belongs to. If  blank, it belongs to
              the Boot Set itself, which only applies to the GenericMetadata type.
            pattern: "(?i)shutdown|boot|configure"
            type: string
          data:
            $ref: '#/components/schemas/NodeChangeList'
    UpdateRequestNodeErrorsList:
      description: |
        This is the payload sent during an update request. It contains
        updates to which errors have occurred and which nodes encountered those errors
      type: array
      items:
        type: object
        properties:
          update_type:
            description: The type of update data 
            pattern: "NodeErrorsList"
            type: string
          phase:
            description: |
              The phase that this data belongs to. If  blank, it belongs to
              the Boot Set itself, which only applies to the GenericMetadata type.
            pattern: "(?i)shutdown|boot|configure"
            type: string
          data:
            $ref: '#/components/schemas/NodeErrorsList'
    UpdateRequestGenericMetadata:
      description: |
        This is the payload sent during an update request. It contains
        updates to metadata, specifically start and stop times
      type: array
      items:
        type: object
        properties:
          update_type:
            description: The type of update data 
            pattern: "GenericMetadata"
            type: string
          phase:
            description: |
              The phase that this data belongs to. If the phase is boot_set, it belongs to
              the Boot Set itself, which only applies to the GenericMetadata type.
            pattern: '(?i)shutdown|boot|configure|boot_set'
            type: string
          data:
            $ref: '#/components/schemas/GenericMetadata'
    ProblemDetails:
      description: An error response for RFC 7807 problem details.
      type: object
      properties:
        type:
          description:
            Relative URI reference to the type of problem which includes human
            readable documentation.
          type: string
          format: uri
          default: "about:blank"
        title:
          description:
            Short, human-readable summary of the problem, should not change by
            occurrence.
          type: string
        status:
          description: HTTP status code
          type: integer
          example: 400
        instance:
          description: A relative URI reference that identifies the specific
            occurrence of the problem
          format: uri
          type: string
        detail:
          description:
            A human-readable explanation specific to this occurrence of the
            problem. Focus on helping correct the problem, rather than giving
            debugging information.
          type: string
      additionalProperties: false
  responses:
    badRequest:
      description: Bad Request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    v1SessionDetails:
      description: Session details
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Session'
    v1SessionStatus:
      description: A list of Boot Set Statuses and metadata
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SessionStatus'
    v1SessionTemplateDetails:
      description: Session template details
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SessionTemplate'
    sessionTemplateNotFound:
      description: The session template was not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    sessionNotFound:
      description: The session was not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
paths:
...
  /v1/sessiontemplate/{session_template_id}:
    parameters:
      - name: session_template_id
        in: path
        description: Session Template ID
        required: true
        schema:
          type: string
    get:
      summary: Get session template by id
      description: |
        Get session template by session_template_id.
        The session_template_id corresponds to the *name*
        of the session template.
      tags:
        - sessiontemplate
      x-openapi-router-controller: bos.controllers.v1.sessiontemplate
      operationId: get_v1_sessiontemplate
      responses:
        200:
          $ref: '#/components/responses/v1SessionTemplateDetails'
        404:
          $ref: '#/components/responses/sessionTemplateNotFound'
    delete:
      summary: Delete a session template
      description: Delete a session template.
      tags:
        - sessiontemplate
      x-openapi-router-controller: bos.controllers.v1.sessiontemplate
      operationId: delete_v1_sessiontemplate
      responses:
        204:
          description: The session template was deleted.
        404:
          $ref: '#/components/responses/sessionTemplateNotFound'
  /v1/session:
    post:
      summary: Create a session
      description: |
        The creation of a session performs the operation
        specified in the SessionCreateRequest
        on the boot set(s) defined in the session template.
      tags:
        - session
      x-openapi-router-controller: bos.controllers.v1.session
      operationId: create_v1_session
      requestBody:
         description: A JSON object for creating a Session
         required: true
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/Session'
      responses:
        200:
          $ref: '#/components/responses/v1SessionDetails'
        400:
          $ref: '#/components/responses/badRequest'

    get:
      summary: List sessions
      description: |
        List all sessions, including those in progress and those complete.
      tags:
        - session
      x-openapi-router-controller: bos.controllers.v1.session
      operationId: get_v1_sessions
      responses:
        200:
          description: A collection of Sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
  /v1/session/{session_id}:
    get:
      summary: Get session details by id
      description: Get session details by session_id.
      tags:
        - session
      x-openapi-router-controller: bos.controllers.v1.session
      operationId: get_v1_session
      responses:
        200:
          $ref: '#/components/responses/v1SessionDetails'
        404:
          $ref: '#/components/responses/sessionNotFound'
    delete:
      summary: Delete session by id
      description: Delete session by session_id.
      tags:
        - session
      x-openapi-router-controller: bos.controllers.v1.session
      operationId: delete_v1_session
      responses:
        204:
          description: The session was deleted.
        404:
          $ref: '#/components/responses/sessionNotFound'

    parameters:
      - name: session_id
        in: path
        description: Session ID
        required: true
        schema:
          type: string
  /v1/session/{session_id}/status:
    parameters:
      - name: session_id
        in: path
        description: Session ID
        required: true
        schema:
          type: string
    get:
      summary: A list of the statuses for the different boot sets.
      description: |
        A list of the statuses for the different boot sets.
      tags:
        - session
      x-openapi-router-controller: bos.controllers.v1.status
      operationId: get_v1_session_status
      responses:
        200:
          description: |
            A collection of boot sets associated with this session.

            The boot set name can be used to query for the status.
          $ref: '#/components/responses/v1SessionStatus'
        404:
          $ref: '#/components/responses/sessionNotFound'
    post:
      summary: Create the initial session status
      description: |
        Creates the initial session status.
      tags:
        - session
      x-openapi-router-controller: bos.controllers.v1.status
      operationId: create_v1_session_status
      requestBody:
         description: A JSON object for creating the status for a session
         required: true
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/SessionStatus'
      responses:
        204:
          $ref: '#/components/responses/v1SessionStatus'
        400:
          $ref: '#/components/responses/badRequest'
    patch:
      summary: Update the session status
      description: |
        Update the session status. You can update the start or stop times.
      tags:
        - session
      x-openapi-router-controller: bos.controllers.v1.status
      operationId: update_v1_session_status
      requestBody:
        description: A JSON object for updating the status for a session
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericMetadata'
      responses:
        200:
          $ref: '#/components/responses/v1SessionStatus'
        404:
          $ref: '#/components/responses/badRequest'
    delete:
      summary: Delete the session status
      description: |
        Deletes an existing Session status
      tags:
        - session
      x-openapi-router-controller: bos.controllers.v1.status
      operationId: delete_v1_session_status
      responses:
        204:
          description: The status was deleted successfully.
        400:
          $ref: '#/components/responses/badRequest'
  /v1/session/{session_id}/status/{boot_set_name}:
    parameters:
      - name: session_id
        in: path
        description: Session ID
        required: true
        schema:
          type: string
      - name: boot_set_name
        in: path
        description: Boot set name
        required: true
        schema:
          type: string
    get:
      summary: Get the status for a boot set.
      description: Get the status for a boot set.
      tags:
        - session
      x-openapi-router-controller: bos.controllers.v1.status
      operationId: get_v1_session_status_by_bootset
      responses:
        200:
          description: A list of the Phase Statuses for the Boot Set and metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BootSetStatus' 
        404:
          $ref: '#/components/responses/sessionNotFound'
    post:
      summary: Create a Boot Set Status
      description: |
        Create a status for a Boot Set
      tags:
        - session
        - status
      x-openapi-router-controller: bos.controllers.v1.status
      operationId: create_v1_boot_set_status
      requestBody:
        description: A JSON object for creating a status for a Boot Set
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BootSetStatus'
      responses:
        201:
          description: The created Boot Set status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BootSetStatus' 
    patch:
      summary: Update the status.
      description: |
        This will change the status for one or more nodes within
        the boot set.
      tags:
        - session
      x-openapi-router-controller: bos.controllers.v1.status
      operationId: update_v1_session_status_by_bootset
      requestBody:
         description: A JSON object for updating the status for a session
         required: true
         content:
           application/json:
             schema:
               anyOf:
                 - $ref: '#/components/schemas/UpdateRequestNodeChangeList'
                 - $ref: '#/components/schemas/UpdateRequestNodeErrorsList'
                 - $ref: '#/components/schemas/UpdateRequestGenericMetadata'
      responses:
        200:
          description: A list of Boot Set Statuses and metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BootSetStatus'
        404:
          $ref: '#/components/responses/sessionNotFound'
    delete:
      summary: Delete the Boot Set status
      description: |
        Deletes an existing Boot Set status
      tags:
        - session
      x-openapi-router-controller: bos.controllers.v1.status
      operationId: delete_v1_boot_set_status
      responses:
        204:
          description: The status was deleted successfully.
        400:
          $ref: '#/components/responses/badRequest'
  /v1/session/{session_id}/status/{boot_set_name}/{phase_name}:
    parameters:
      - name: session_id
        in: path
        description: Session ID
        required: true
        schema:
          type: string
      - name: boot_set_name
        in: path
        description: Boot set name
        required: true
        schema:
          type: string
      - name: phase_name
        in: path
        description: The phase name
        required: true
        schema:
          type: string
    get:
      summary: Get the status for a specific boot set and phase.
      description: Get the status for a specific boot set and phase.
      tags:
        - session
      x-openapi-router-controller: bos.controllers.v1.status
      operationId: get_v1_session_status_by_bootset_and_phase
      responses:
        200:
          description: A list of the nodes in the Phase and Category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhaseStatus' 
        404:
          $ref: '#/components/responses/sessionNotFound'
...